refactor(trimming): implement pre-trimming
This commit refactors trimming such that the staking-miner is
responsible to ensure that trimming doesn't have to be done be on-chain.

Instead the staking-miner iteratively removes the votes by one voter at
the time and re-computes the solution. Once a solution with no trimming
is found that solution is submitted to the chain