static config: Add support for AH and staking-async chains (#1064)
* static config: Add support for AH and staking-async chains

- Deprecated `asset-hub-next`, replaced by `staking-async-parachain` and
`staking-async-rc` in
https://github.com/paritytech/polkadot-sdk/pull/8422
- Add support for statemint, statemine, westmint AH chains, without
dropping the polkadot / kusama / westend RC variants until we deprecate
the legacy `monitor` path and rely only on EPMB

For the time being, static miner's config for StakingAsync is a
copy-paste of Westend one but will evolve independently.

As a separate topic (see #994), a partial / complete removal of the
static configuration per chain will be tackled.

* Remove staking-async support from legacy

This change removes staking-async runtime support from the legacy monitor
code path, replacing it with appropriate error messages for unsupported
cases.