[EXPERIMENTAL - DO NOT MERGE] Dummy miner only listening to new blocks
To investigate the issue of the stalling subscription (see #1119,
doesn't actually try to mine and clean-up.

The idea is to run it ideally on the same runner which runs the real
miner on WAH and observes if it stalls in this ultra-simplified version.

If it does, we can safely excluse any side effect coming from the
janitor task or the miner task itself.